Fixes to use the local copy of github.com/googl/fhir/go module.
Later versions of fhir/go break the current code in this repo,
trying to only point to GH repo with a tag version caused
errors due to the size fo the GH repo, so I copied the module source
at tag v0.7.0 and COPYed it to the Docker build and mod.go "replace"ed
to point to this local copy.

A "docker-compose build --no-cache" at the repo root
should succeed now but this doesn't
necessarily mean the images are functional. There are many warnings
related to version resolutions and conflicts (I think mostly Python
related) and this could be a problem. I think there are at least few
unpinned library dependencies and the version resolutions now might
be a problem compared to how they were resolved in Jan 2021, as in
the official Google provided images here:

http://gcr.io/cloud-healthcare-containers

We can dig into the Google images if we need to find which versions
were resolved at the time if that becomes necessary.